[feat] wires up rendering, kinda closing the loop on everything
