+++ src/components/editor/array-editor.tsx
@@ -1,6 +1,7 @@
-import React from "react";
+import React, { useRef } from "react";
 import ArraySelector from "../selector/array-selector";
 import { ZarrView } from "@/state";
+import ArrayRenderer from "./array-renderer";
 
 const margin = `1em`;
 
@@ -16,6 +17,7 @@ export default function ArrayEditor({
   viewerIdx,
   sidebarWidth,
 }: IArrayEditorProps) {
+  const parentRef = useRef<HTMLDivElement>(null);
   return (
     <div
       className="h-screen relative"
@@ -23,6 +25,7 @@ export default function ArrayEditor({
     >
       <div className="absolute w-full h-full top-0 left-0 ">
         <div
+          ref={parentRef}
           className="relative my-2 border-2 border-gray-300 rounded-xl bg-white"
           style={{
             margin: `${margin}`,
@@ -31,6 +34,8 @@ export default function ArrayEditor({
         >
           {/* this is where the rendering component should go, with a ref to its parent container so it can set its size properly. */}
 
+          <ArrayRenderer viewer={viewer} parentElement={parentRef} />
+
           <ArraySelector
             viewer={viewer}
             viewerIdx={viewerIdx}
@@ -38,10 +43,7 @@ export default function ArrayEditor({
             className="absolute top-[50%] left-[50%]"
             style={{ transform: `translate(-50%,-50%)` }}
           />
-
-          {/* This is where any additional controls should go. */}
         </div>
-        {/* This is where the text-editor pane should pop out */}
       </div>
     </div>
   );
